kconfig: refresh all kconfigs to reflect latest code
Previously the kconfigs from openeuler_defconfig are missing some
options. This will cause the RPM build process to fail as it expects no
supprise from kconfig change during build process. To work around this
we have to pass the "-nf" parameter to the tar-up.sh script every
time. And now it is fixed.

The kconfigs are refreshed by following the steps in README.

Changes are:

  * All driver and hardware support options are set to M or Y (if no M).
  * All other options are set to their upstream defaults.

Signed-off-by: Kai Liu <kai.liu@suse.com>
Change-Id: Iddd0136be4cc6050903d9a3a8074921553a12290