SEL-2.0: port patches for ACPI per the report of git-fixes
Per git-fixes, the two commit are needed.

935dff305da2 ACPI: CPPC: Add NULL pointer check to cppc_get_perf()
0df316b8ec04 ACPI: IORT: Handle device properties with software node API

And to make compiler happy, the relevant patch is ported as well.

151f6ff78cd software node: Provide replacement for device_add_properties()

then there is a fix for above with slightly tweak due to the difference
of software_node_notify ini 5.10 kernel.

5aeb05b27f software node: balance refcount for managed software nodes

Also 5864-ACPI-IORT-Enable-stall-support-for-platform-devices.patch is
refreshed because of the new added patches.

Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>
Change-Id: Ieced0c9a32edd82ca6bde635e9f13bda15537cf8