xfs: remove callback dequeue loop from
From bbe4d31d85e2c44ff4f9c1809f2a3aa44a3519e1 Mon Sep 17 00:00:00 2001
 xlog_state_do_iclog_callbacks
Git-commit: b6903358c230c517b29ecdb6123276d96cc0beab
Patch-mainline: v5.14-rc1
References: bsn#12
Modified-by-SEL: No

If we are processing callbacks on an iclog, nothing can be
concurrently adding callbacks to the loop. We only add callbacks to
the iclog when they are in ACTIVE or WANT_SYNC state, and we
explicitly do not add callbacks if the iclog is already in IOERROR
state.

The only way to have a dequeue racing with an enqueue is to be
processing a shutdown without a direct reference to an iclog in
ACTIVE or WANT_SYNC state. As the enqueue avoids this race
condition, we only ever need a single dequeue operation in
xlog_state_do_iclog_callbacks(). Hence we can remove the loop.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.14-rc1-xfs-remove-callback-dequeue-loop-from-xlog_state_do_.patch